fix: move timeslots-api to booking-api
