+++ services/booking-api/src/helpers/createSlots.js
