+++ services/booking-api/src/helpers/getTimeSpans.js
