+++ services/booking-api/src/lambdas/getTimeSlots.js
