+++ services/booking-api/test/helpers/createSlots.test.js
