+++ services/booking-api/test/lambdas/getTimeSlots.test.js
