+++ services/booking-api/test/lambdas/update.test.ts
@@ -4,12 +4,12 @@ const messages = require('@helsingborg-stad/npm-api-error-handling/assets/errorM
 import { main } from '../../src/lambdas/update';
 import booking from '../../src/helpers/booking';
 import { isTimeslotTaken } from '../../src/helpers/isTimeslotTaken';
-import getTimeSpans from '../../src/libs/getTimeSpans';
+import getTimeSpans from '../../src/helpers/getTimeSpans';
 import { areAllAttendeesAvailable } from '../../src/helpers/timeSpanHelper';
 
 jest.mock('../../src/helpers/booking');
 jest.mock('../../src/helpers/isTimeslotTaken');
-jest.mock('../../src/libs/getTimeSpans');
+jest.mock('../../src/helpers/getTimeSpans');
 jest.mock('../../src/helpers/timeSpanHelper');
 
 const { search, create, cancel } = jest.mocked(booking);
