--- services/timeslots-api/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "timeslots-api",
-  "version": "0.0.1",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "cd ../../ && node node_modules/.bin/jest timeslots-api"
-  },
-  "keywords": [],
-  "author": "Helsingborg Stad",
-  "license": "MIT"
-}
