--- services/timeslots-api/serverless.yml
@@ -1,50 +1,0 @@
-service: timeslots-api
-
-plugins:
-  - serverless-bundle
-  - serverless-add-api-key
-
-projectDir: ../../
-
-custom: ${file(../../serverless.common.yml):custom}
-
-package:
-  individually: true
-
-provider:
-  lambdaHashingVersion: 20201221
-  name: aws
-  runtime: nodejs14.x
-  stage: dev
-  region: eu-north-1
-  endpointType: regional
-  tracing:
-    lambda: true
-  apiGateway:
-    restApiId: !ImportValue ${self:custom.stage}-ExtApiGatewayRestApiId
-    restApiRootResourceId: !ImportValue ${self:custom.stage}-ExtApiGatewayRestApiRootResourceId
-    restApiResources:
-      timeslots: !ImportValue ${self:custom.stage}-ExtApiGatewayResourceTimeSlots
-
-  environment:
-    stage: ${self:custom.stage}
-    resourcesStage: ${self:custom.resourcesStage}
-
-  iam:
-    role:
-      statements:
-        - Effect: Allow
-          Action:
-            - ssm:GetParameter
-          Resource:
-            - !Sub arn:aws:ssm:${self:provider.region}:${AWS::AccountId}:parameter/datatorgetEnvs/${self:provider.stage}
-
-functions:
-  getTimeSlots:
-    handler: src/lambdas/getTimeSlots.main
-    events:
-      - http:
-          path: timeslots/getTimeSlots
-          method: post
-          private: true
-          cors: true
