--- services/timeslots-api/src/libs
@@ -1,1 +1,0 @@
-../../../libs
