--- services/timeslots-api/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "baseUrl": "src",
-    "rootDir": "./src"
-  },
-  "exclude": [],
-  "include": ["src/**/*"]
-}
